requirements.txt
altair==5.4.0
asttokens==2.4.1
attrs==24.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
contourpy==1.3.0
cycler==0.12.1
debugpy==1.8.5
decorator==5.1.1
executing==2.0.1
fonttools==4.53.1
gitdb==4.0.11
GitPython==3.1.43
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.9.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
narwhals==1.5.2
nest-asyncio==1.6.0
nltk==3.9.1
numpy==2.1.0
-e git+https://github.com/naufal-yafi/model.text-mining.git@cffef2e8d7e33ac4a3a88064e8b5e44e62912eb8#egg=package
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
pyzmq==26.1.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rfc3986==1.5.0
rich==13.7.1
rpds-py==0.20.0
Sastrawi==1.0.1
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
setuptools==75.3.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
stack-data==0.6.3
streamlit==1.37.1
tenacity==8.5.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.2
wcwidth==0.2.13
